Extract managed Ghostty appearance config store
The appearance settings flow now delegates managed Ghostty config persistence to a dedicated store. The store owns color resolution, marker block replacement, serialized live writes, and read failure handling so AppearanceSettings remains an app-appearance coordinator instead of an on-disk config writer.

Constraint: Review feedback required separating persistence from AppearanceSettings and preserving existing config on non-missing read failures.

Rejected: Keep helpers in AppearanceSettings | it mixed appearance mode coordination with config parsing and file I/O.

Rejected: Treat any read failure as empty config | invalid or unreadable user config could be overwritten.

Confidence: medium

Scope-risk: moderate

Directive: Keep managed Ghostty config parsing and file I/O in ManagedGhosttyAppearanceConfigStore.

Tested: swiftc -parse Sources/AppearanceSettings.swift; swiftc -parse Sources/ManagedGhosttyAppearanceConfigStore.swift; swiftc -parse Sources/GhosttyTerminalView.swift; swiftc -parse cmuxTests/AppearanceSettingsTests.swift; git diff --check

Not-tested: Local XCTest execution per repository policy; full validation delegated to CI
