+++ Sources/ManagedGhosttyAppearanceConfigStore.swift
@@ -1,0 +1,191 @@
+import AppKit
+import Foundation
+
+struct ManagedGhosttyAppearanceConfigStore {
+    let environment: ConfigSourceEnvironment
+    let defaults: UserDefaults
+
+    init(
+        environment: ConfigSourceEnvironment = .live(),
+        defaults: UserDefaults = .standard
+    ) {
+        self.environment = environment
+        self.defaults = defaults
+    }
+
+    func persistManagedTerminalAppearanceConfig(
+        mode: AppearanceMode,
+        appAppearance: NSAppearance?,
+        source: String
+    ) {
+        let normalized = AppearanceSettings.mode(for: mode.rawValue)
+        let colorScheme = Self.managedTerminalColorScheme(
+            for: normalized,
+            appAppearance: appAppearance,
+            defaults: defaults
+        )
+        let managedBlock = Self.managedTerminalAppearanceBlock(
+            mode: normalized,
+            colorScheme: colorScheme
+        )
+
+        Self.persistManagedTerminalAppearanceBlock(
+            managedBlock,
+            source: source,
+            environment: environment
+        )
+    }
+
+    static func livePersistenceTask(
+        mode: AppearanceMode,
+        appAppearance: NSAppearance?,
+        defaults: UserDefaults,
+        source: String
+    ) -> Task<Void, Never>? {
+        let normalized = AppearanceSettings.mode(for: mode.rawValue)
+        let colorScheme = managedTerminalColorScheme(
+            for: normalized,
+            appAppearance: appAppearance,
+            defaults: defaults
+        )
+        let managedBlock = managedTerminalAppearanceBlock(
+            mode: normalized,
+            colorScheme: colorScheme
+        )
+
+        let completion = AsyncStream<Void> { continuation in
+            managedTerminalAppearanceWriteQueue.async {
+                persistManagedTerminalAppearanceBlock(
+                    managedBlock,
+                    source: source
+                )
+                continuation.yield(())
+                continuation.finish()
+            }
+        }
+
+        return Task(priority: .utility) {
+            for await _ in completion {}
+        }
+    }
+
+    private static let managedTerminalAppearanceWriteQueue = DispatchQueue(
+        label: "com.cmux.appearance.managedTerminalConfig"
+    )
+
+    private static func managedTerminalColorScheme(
+        for mode: AppearanceMode,
+        appAppearance: NSAppearance?,
+        defaults: UserDefaults
+    ) -> GhosttyConfig.ColorSchemePreference {
+        switch mode {
+        case .light:
+            return .light
+        case .dark:
+            return .dark
+        case .system, .auto:
+            return GhosttyConfig.currentColorSchemePreference(
+                appAppearance: appAppearance ?? AppearanceSettings.systemNSAppearance(defaults: defaults),
+                defaults: defaults
+            )
+        }
+    }
+
+    private static let managedTerminalAppearanceBeginMarker = "# cmux-managed-appearance: begin"
+    private static let managedTerminalAppearanceEndMarker = "# cmux-managed-appearance: end"
+
+    private static func managedTerminalAppearanceBlock(
+        mode: AppearanceMode,
+        colorScheme: GhosttyConfig.ColorSchemePreference
+    ) -> String {
+        let body = GhosttyConfig.cmuxDefaultThemeConfigContents(
+            preferredColorScheme: colorScheme
+        )
+        .trimmingCharacters(in: .whitespacesAndNewlines)
+
+        return """
+        \(managedTerminalAppearanceBeginMarker)
+        # Generated by cmux from the app appearance setting.
+        # mode = \(mode.rawValue)
+        \(body)
+        \(managedTerminalAppearanceEndMarker)
+
+        """
+    }
+
+#if compiler(>=6.2)
+    @concurrent
+#endif
+    private nonisolated static func persistManagedTerminalAppearanceBlock(
+        _ managedBlock: String,
+        source: String,
+        environment: ConfigSourceEnvironment = .live()
+    ) {
+        do {
+            let url = environment.cmuxConfigURL
+            let existingContents = try existingContents(at: url, environment: environment)
+            let updatedContents = replacingManagedTerminalAppearanceBlock(
+                in: existingContents,
+                with: managedBlock
+            )
+            guard updatedContents != existingContents else { return }
+            try environment.writeCmuxConfigContents(updatedContents)
+        } catch {
+            #if DEBUG
+            cmuxDebugLog("appearance.ghosttyConfig.persist.failed source=\(source) error=\(error)")
+            #endif
+        }
+    }
+
+    private nonisolated static func existingContents(
+        at url: URL,
+        environment: ConfigSourceEnvironment
+    ) throws -> String {
+        do {
+            return try String(contentsOf: url, encoding: .utf8)
+        } catch {
+            let nsError = error as NSError
+            if nsError.domain == NSCocoaErrorDomain,
+               nsError.code == NSFileReadNoSuchFileError {
+                return ""
+            }
+            if !environment.fileManager.fileExists(atPath: url.path) {
+                return ""
+            }
+            throw error
+        }
+    }
+
+    private nonisolated static func replacingManagedTerminalAppearanceBlock(
+        in contents: String,
+        with managedBlock: String
+    ) -> String {
+        if contents.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
+            return managedBlock
+        }
+
+        if let startRange = contents.range(of: managedTerminalAppearanceBeginMarker) {
+            let replacementStart = contents[..<startRange.lowerBound]
+                .lastIndex(of: "\n")
+                .map { contents.index(after: $0) } ?? contents.startIndex
+            guard let endRange = contents.range(
+                of: managedTerminalAppearanceEndMarker,
+                range: startRange.upperBound..<contents.endIndex
+            ) else {
+                var updated = contents
+                updated.replaceSubrange(replacementStart..<contents.endIndex, with: managedBlock)
+                return updated.hasSuffix("\n") ? updated : updated + "\n"
+            }
+
+            let replacementEnd = contents[endRange.upperBound...]
+                .firstIndex(of: "\n")
+                .map { contents.index(after: $0) } ?? contents.endIndex
+            var updated = contents
+            updated.replaceSubrange(replacementStart..<replacementEnd, with: managedBlock)
+            return updated.hasSuffix("\n") ? updated : updated + "\n"
+        }
+
+        let separator = contents.hasSuffix("\n") ? "\n" : "\n\n"
+        return contents + separator + managedBlock
+    }
+}
